requirements.txt
BeautifulSoup==3.2.0
CherryPy==3.2.0
Django==1.3
Fabric==1.0.0
MySQL-python==1.2.3
PIL==1.1.7
PyYAML==3.09
South==0.7.3
Whoosh==1.8.2
akismet==0.2.0
boto==2.0b4
distribute==0.6.14
django-annoying==0.7.6
django-autocomplete==1.0.dev49
django-backcap==0.0.0
django-batch-select==0.2.2
django-blog-zinnia==0.8
django-debug-toolbar==0.8.4
django-exceptional==0.1.3
django-fabtastic==1.1
django-guardian==1.0.2
django-haystack==1.1.0
django-imagekit==0.3.5
django-indexer==0.2.1
django-keyedcache==1.4-1
django-livesettings==1.4-7
django-mptt==0.4.2
django-notification==0.1.5
django-paging==0.2.3.1
-e hg+http://bitbucket.org/ubernostrum/django-registration/@58eef8330b0f23da4645eb01c078621eb3998f39#egg=django_registration-dev
django-sentry==1.6.10
django-tagging==0.3.1
django-tastypie==0.9.9
django-templatetag-sugar==0.1
django-userena==1.0.0
django-voting==0.1
django-xmlrpc==0.1.3
docutils==0.7
easy-thumbnails==1.0-alpha-17
gunicorn==0.12.0
httplib2==0.6.0
ipdb==0.3
ipython==0.10.1
mimeparse==0.1.3
numpy==1.5.1
oauth2==1.5.169
paramiko==1.7.6
-e git://gitorious.org/pyclean/mainline.git@01ac05c3ba54b5d9b487e8037cffb41903c17bef#egg=pyclean-dev
pycrypto==2.3
pydot==1.0.4
pyparsing==1.5.5
python-dateutil==1.5
python-digest==1.7
python-graph-core==1.8.0
python-graph-dot==1.8.0
pyutf8==0.1.1
setproctitle==1.1.1
silk-config==0.1
silk-deployment==0.2.5
sorl-thumbnail==11.01
uuid==1.30
vobject==0.8.1c
wsgiref==0.1.2
yolk==0.4.1
